move todo state setting prsing a document function.
